The training observation number and update frequency did not reflect the number of classes in the data.
CURRENT ISSUE: We have 10080 observations, we train on 2880 ov them, so our faultChecks_xts matrix has 7200 rows as expected. The number of alarmed observations identified is 83, so the non-alarmed data matrix should have close to 7200 - 83. However, it has 4195.
FUTHER ISSUE: I found earlier that an observation, the second flag in a four-flag series, should have been marked as an alarm but wasn't. No idea there.